+++ open-sphere-base/core/src/main/java/io/opensphere/core/pipeline/renderer/buffered/MultiDrawIndexBufferObject.java
@@ -78,8 +78,8 @@ public class MultiDrawIndexBufferObject implements BufferObject
         }
         else
         {
-            rc.getGL2().glMultiDrawElements(drawMode, (IntBuffer)mySizeBuffer.rewind(), GL.GL_UNSIGNED_INT,
-                    myPointerBuffer.rewind(), myPointerBuffer.limit());
+            rc.getGL2().glMultiDrawElements(drawMode, mySizeBuffer.rewind(), GL.GL_UNSIGNED_INT, myPointerBuffer.rewind(),
+                    myPointerBuffer.limit());
             return true;
         }
     }
